+++ glutin-winit/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- **Breaking:** Update _winit_ to `0.29.2`. See [winit's CHANGELOG](https://github.com/rust-windowing/winit/releases/tag/v0.29.2) for more info.
+
 # Version 0.4.1-beta
 
 - **Breaking:** Update _winit_ to `0.29.1-beta`. See [winit's CHANGELOG](https://github.com/rust-windowing/winit/releases/tag/v0.29.1-beta) for more info.